+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - clojure: conjure, paredit, parinfer, parpar & nvim-treesitter-sexp plugins
 - neo-tree: override examples, show hidden in alternate style
 - plantuml: key mapping to call out to plantuml jar or binary
+- example-config: add new plugins & override existing plugin config 
 
 ## Changed
 - readme: expand on description and use